+++ LCUD_Project/Cuboid.h
@@ -6,13 +6,13 @@
 class Cuboid : public Container{
     //attribute dimensions
     private:
-        int Cwidth,Clength,Cheigth;
+        float Cwidth,Clength,Cheigth;
     public:
-        Cuboid(int width,int length,int heigth);
+        Cuboid(float width,float length,float heigth);
         ~Cuboid(){;}
         std::string to_string();
         float Cuboid::getCuboidWidth();
         float Cuboid::getCuboidLength();
         float Cuboid::getCuboidHeight();
-};
+}; 
 #endif
